Renamed entities.js to Entities.js
